Added confirmation prompt to import site task
git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2572 c89fab08-2f9b-42bf-8176-bd3ad6bee547

+++ lib/task/aImportSiteTask.class.php
@@ -44,10 +44,13 @@ EOF;
 
   protected function execute($args = array(), $options = array())
   {
-    $conn = 'doctrine';
-    
     $databaseManager = new sfDatabaseManager($this->configuration);
     $connection = $databaseManager->getDatabase($options['connection'])->getDoctrineConnection();
+
+    if(!$this->askConfirmation("Importing any content will erase any existing content, are you sure? [y/N]", 'QUESTION_LARGE', false))
+    {
+      die("Import CANCELLED.  No changes made.\n");
+    }
 
     $connection->dropDatabase();
     $connection->createDatabase();
